+++ CotEditor/Sources/MultipleReplacementListViewController.swift
@@ -154,7 +154,11 @@ final class MultipleReplacementListViewController: NSViewController, NSMenuItemV
         
         guard let settingName = self.targetSettingName(for: sender) else { return }
         
-        try? ReplacementManager.shared.duplicateSetting(name: settingName)
+        do {
+            try ReplacementManager.shared.duplicateSetting(name: settingName)
+        } catch {
+            self.presentError(error)
+        }
     }
     
     
@@ -189,12 +193,16 @@ final class MultipleReplacementListViewController: NSViewController, NSMenuItemV
         savePanel.canSelectHiddenExtension = true
         savePanel.nameFieldLabel = "Export As:".localized
         savePanel.nameFieldStringValue = settingName
-        savePanel.allowedFileTypes = []
+        savePanel.allowedFileTypes = ReplacementManager.shared.filePathExtensions
         
         savePanel.beginSheetModal(for: self.view.window!) { (result: NSApplication.ModalResponse) in
             guard result == .OK else { return }
             
-            try? ReplacementManager.shared.exportSetting(name: settingName, to: savePanel.url!)
+            do {
+                try ReplacementManager.shared.exportSetting(name: settingName, to: savePanel.url!, hidesExtension: savePanel.isExtensionHidden)
+            } catch {
+                self.presentError(error)
+            }
         }
     }