+++ CHANGELOG.md
@@ -12,6 +12,7 @@ Change Log
 
 ### Fixes
 
+- Fix an issue where the multi replacement feature exported definitions without its file extension.
 - Fix an issue where an unhidable empty progress dialog was displayed if another dialog was displayed when the task started.
 - Address an issue where the application could rarely crash during syntax highlighting.